development completed

+++ src/main/java/com/dolap/repository/IProductDao.java
@@ -11,7 +11,7 @@ import com.dolap.entity.Product;
  * @author umutates
  *2 Şub 2018
  */
-public interface IProductDao extends IBaseDao<Product,Long>{
+public interface IProductDao extends IBaseDao<Product,Integer>{
 
 	public List<Product> findAll();
